Add gem dependencies
To product XML, I'm going to use nokogiri. It's a well-known library
with broad support and great interface. If performance is essential, I
can switch over to ox.

I'm going to use rake, as it provides excellent file-related utilities
and simple ways to work with files in parallel.

RSpec is my favourite testing framework.